+++ example/lib/cloud_firestore_api/data/dtos/example_dto.dart
@@ -1,0 +1,23 @@
+class ExampleDTO {
+  ExampleDTO({
+    required this.thisIsAString,
+    required this.thisIsANumber,
+    required this.thisIsABoolean,
+  });
+
+  final String thisIsAString;
+  final double thisIsANumber;
+  final bool thisIsABoolean;
+
+  factory ExampleDTO.fromJson(Map<String, dynamic> json) => ExampleDTO(
+        thisIsAString: json["thisIsAString"],
+        thisIsANumber: json["thisIsANumber"],
+        thisIsABoolean: json["thisIsABoolean"],
+      );
+
+  Map<String, dynamic> toJson() => {
+        "thisIsAString": thisIsAString,
+        "thisIsANumber": thisIsANumber,
+        "thisIsABoolean": thisIsABoolean,
+      };
+}